Add files via upload
This repository contains the excel file (NSQD+.xlsx) which has data from the National Stormwater Quality Database (NSQD) and more recent sources as well as two R files. These R files were used to analyze the NSQD+ in section 3.1 of the manuscript and create and analyze the NSQD++ in section 3.2 of the manuscript.